fix(close-button): keep static color and click handling self-contained
Revert close-button-driven changes from ButtonBase/Button and keep
static-color behavior on CloseButton so this migration is scoped and
independent from shared action-button/base coordination.
