+++ 2nd-gen/packages/swc/components/button/Button.ts
@@ -20,6 +20,7 @@ import {
   BUTTON_VARIANTS,
   ButtonBase,
   type ButtonFillStyle,
+  type ButtonStaticColor,
   type ButtonVariant,
 } from '@spectrum-web-components/core/components/button';
 
@@ -93,6 +94,13 @@ export class Button extends ButtonBase {
    */
   @property({ type: String, reflect: true, attribute: 'fill-style' })
   public fillStyle: ButtonFillStyle = 'fill';
+
+  /**
+   * Static color treatment for display over colored or image backgrounds.
+   * Only supported with `primary` and `secondary` variants.
+   */
+  @property({ type: String, reflect: true, attribute: 'static-color' })
+  public staticColor?: ButtonStaticColor;
 
   /**
    * Whether overflowing text is truncated with an ellipsis rather than
